+++ src/main/java/com/tchepannou/kiosk/ranker/ScoreProvider.java
@@ -1,5 +1,5 @@
 package com.tchepannou.kiosk.ranker;
 
 public interface ScoreProvider {
-    int get(Rankable rankable);
+    long get(Rankable rankable);
 }
