+++ src/main/java/com/tchepannou/kiosk/ranker/score/ImageScoreProvider.java
@@ -5,7 +5,7 @@ import com.tchepannou.kiosk.ranker.ScoreProvider;
 
 public class ImageScoreProvider implements ScoreProvider {
     @Override
-    public int get(final Rankable rankable) {
+    public long get(final Rankable rankable) {
         return rankable.isWithImage() ? 100 : 0;
     }
 }
